4.21*: Blog List Expansion, step 9 can only delete own blog

+++ part4/blog/controllers/blogs.js
@@ -68,7 +68,17 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   const id = request.params.id
+  const blog = await Blog.findById(id)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'not permitted to delete non-owned resource'})
+  }
   await Blog.findByIdAndDelete(id)
   response.status(204).end()
 })
